reworked where things are and started suggestion work
