--- main.go
@@ -1,357 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"strings"
-
-	"github.com/charmbracelet/bubbles/help"
-	"github.com/charmbracelet/bubbles/textinput"
-	"github.com/charmbracelet/bubbles/viewport"
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-)
-
-func main() {
-	p := tea.NewProgram(newModel(), tea.WithAltScreen())
-	_, err := p.Run()
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-// custom messages/commands
-type advanceStage struct{}
-
-func nextStage() tea.Msg {
-	return advanceStage{}
-}
-
-// icons
-var (
-	documentIcon   = string([]byte{0xF0, 0x9F, 0x97, 0x8E}) // UTF-8 encoding for "🗎"
-	suggestionIcon = string([]byte{0xE2, 0x91, 0x82})       // UTF-8 encoding for "⑂"
-)
-
-// model
-type model struct {
-	title                       string
-	currentStage, width, height int
-	bodyHeight                  *int
-	stages                      []string
-	bodyContent                 []tea.Model
-	ready                       bool
-	body                        viewport.Model
-	help                        help.Model
-}
-
-func newModel() model {
-	var height int
-	return model{
-		title: "Epub Linter Manually Fixable Issues",
-		// title: "EL MFI",
-		bodyContent: []tea.Model{
-			newSectionBreak(true),
-			newSuggestions(&height),
-		},
-		help:       help.New(),
-		body:       viewport.New(0, 0),
-		bodyHeight: &height,
-		stages: []string{
-			"Section Break",
-			"Suggestions",
-			"CSS File Selection",
-		},
-	}
-}
-
-func (m model) Init() tea.Cmd {
-	return nil
-}
-
-func (m model) View() string {
-	if m.ready {
-		var (
-			header       = m.headerView()
-			footer       = m.footerView()
-			headerHeight = lipgloss.Height(header) + headerBorderStyle.GetBorderBottomSize()
-			footerHeight = lipgloss.Height(footer) + footerBorderStyle.GetBorderTopSize()
-		)
-
-		m.body.Width = m.width
-		m.body.Height = max(0, m.height-(headerHeight+footerHeight)+2)
-		// preserve pointer value to allow for proper seinding of the value to any views that need it
-		// using a new pointer would break the references in the body views
-		(*m.bodyHeight) = m.body.Height
-		m.body.SetContent(m.bodyContent[m.currentStage].View())
-
-		return lipgloss.JoinVertical(lipgloss.Center, header, m.body.View(), footer)
-	}
-
-	return ""
-}
-
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "esc":
-			return m, tea.Quit
-		case "ctrl+c":
-			// TODO: make sure this is an error once ready
-			return m, tea.Quit
-		}
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-		m.ready = true
-	case advanceStage:
-		// TODO: add more in depth logic
-		m.currentStage++
-	}
-
-	var cmd tea.Cmd
-	m.bodyContent[m.currentStage], cmd = m.bodyContent[m.currentStage].Update(msg)
-
-	return m, cmd
-}
-
-func max(a, b int) int {
-	if a > b {
-		return a
-	}
-
-	return b
-}
-
-// styles
-var (
-	titleStyle            = lipgloss.NewStyle().Bold(true).Foreground(lipgloss.Color("#CBACF7"))
-	inactiveStyle         = lipgloss.NewStyle().Faint(true)
-	activeStyle           = lipgloss.NewStyle().Foreground(lipgloss.Color("#fab387"))
-	headerBorderStyle     = lipgloss.NewStyle().Border(lipgloss.RoundedBorder(), true).BorderForeground(lipgloss.Color("12"))
-	controlsStyle         = lipgloss.NewStyle().Faint(true).Bold(true)
-	footerBorderStyle     = lipgloss.NewStyle().Border(lipgloss.RoundedBorder(), true).BorderForeground(lipgloss.Color("12"))
-	fileNameStyle         = lipgloss.NewStyle().Foreground(lipgloss.Color("#a6e3a1"))
-	suggestionNameStyle   = lipgloss.NewStyle().Foreground(lipgloss.Color("#f5e0dc"))
-	leftStatusBorderStyle = lipgloss.NewStyle().Border(lipgloss.RoundedBorder(), true).BorderForeground(lipgloss.Color("12"))
-)
-
-func (m model) headerView() string {
-	return headerBorderStyle.Render(fillLine(lipgloss.JoinHorizontal(lipgloss.Center, titleStyle.Render(m.title), " | ", m.getStageHeaders()), m.width-headerBorderStyle.GetHorizontalBorderSize()))
-}
-
-func (m model) getStageHeaders() string {
-	var stageHeaders = make([]string, len(m.stages))
-
-	var style lipgloss.Style
-	for i, header := range m.stages {
-		if i == m.currentStage {
-			style = activeStyle
-		} else {
-			style = inactiveStyle
-		}
-
-		stageHeaders[i] = style.Render(header)
-	}
-
-	return strings.Join(stageHeaders, inactiveStyle.Render(" > "))
-}
-
-func (m model) footerView() string {
-	var s strings.Builder
-	s.WriteString(fillLine(controlsStyle.Render("Controls:"), m.width-footerBorderStyle.GetHorizontalBorderSize()) + "\n")
-
-	var controls []string
-	switch m.currentStage {
-	case 0:
-		controls = []string{
-			"Enter: Accept",
-			"Esc: Quit",
-			"Ctrl+C: Exit without saving",
-		}
-	case 1:
-		// if m.potentiallyFixableIssuesInfo.isEditing {
-		// 	controls = []string{
-		// 		"Ctrl+R: Reset",
-		// 		"Ctrl+E: Cancel edit",
-		// 		"Ctrl+S: Accept",
-		// 		"Esc: Quit",
-		// 		"Ctrl+C: Exit without saving",
-		// 	}
-		// } else if m.potentiallyFixableIssuesInfo.currentSuggestionState != nil && m.potentiallyFixableIssuesInfo.currentSuggestionState.isAccepted {
-		// controls = []string{
-		// 	"← / → : Previous/Next Suggestion",
-		// 	"C: Copy",
-		// 	"Esc: Quit",
-		// 	"Ctrl+C: Exit without saving",
-		// }
-		// } else {
-		controls = []string{
-			"← / → : Previous/Next Suggestion",
-			"E: Edit",
-			"C: Copy",
-			"Enter: Accept",
-			"Esc: Quit",
-			"Ctrl+C: Exit without saving",
-		}
-		// }
-	case 2:
-		controls = []string{
-			"↑ / ↓ : Previous/Next Suggestion",
-			"Enter: Accept",
-			"Ctrl+C: Exit without saving",
-		}
-	}
-
-	var (
-		line     strings.Builder
-		maxWidth = m.width
-	)
-	for _, help := range controls {
-		if line.Len() == 0 {
-			line.WriteString(help)
-			s.WriteString(help)
-		} else if line.Len()+len(help)+3 <= maxWidth {
-			s.WriteString(" • " + help)
-			line.WriteString(" • " + help)
-		} else {
-			s.WriteString("\n")
-			line.Reset()
-
-			line.WriteString(help)
-			s.WriteString(help)
-		}
-	}
-
-	if line.Len() != 0 {
-		s.WriteString("\n")
-	}
-
-	return footerBorderStyle.Render(s.String())
-}
-
-func fillLine(currentValue string, width int) string {
-	var amountToFill = width - lipgloss.Width(currentValue)
-	if amountToFill < 1 {
-		return currentValue
-	}
-
-	return currentValue + strings.Repeat(" ", amountToFill)
-}
-
-// Section Break
-type sectionBreak struct {
-	input        textinput.Model
-	contextBreak string
-}
-
-func newSectionBreak(focus bool) sectionBreak {
-	ti := textinput.New()
-	ti.Width = 20
-	ti.CharLimit = 200
-	ti.Placeholder = "Section break"
-
-	if focus {
-		ti.Focus()
-	}
-
-	return sectionBreak{
-		input: ti,
-	}
-}
-
-func (m sectionBreak) Init() tea.Cmd {
-	return nil
-}
-
-func (m sectionBreak) View() string {
-	return m.input.View()
-}
-
-func (m sectionBreak) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var (
-		cmds []tea.Cmd
-		cmd  tea.Cmd
-	)
-	m.input, cmd = m.input.Update(msg)
-	cmds = append(cmds, cmd)
-
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "enter":
-			m.contextBreak = strings.TrimSpace(m.input.Value())
-			if m.contextBreak != "" {
-				cmds = append(cmds, nextStage)
-			}
-		}
-	}
-
-	return m, tea.Batch(cmds...)
-}
-
-// Suggestions
-type suggestions struct {
-	height                             *int
-	currentFile, currentSuggestionName string
-	// suggestionData                     []fileSuggestionInfo
-
-	// currentSuggestionIndex, potentialFixableIssueIndex, currentFileIndex int
-	// potentialIssues []potentiallyFixableIssue
-	// currentIssue                                                                  *potentiallyFixableIssue
-	// cssUpdateRequired, addCssSectionBreakIfMissing, addCssPageBreakIfMissing, isEditing bool
-	// currentSuggestionState *suggestionState
-	// suggestionEdit         textarea.Model
-	// suggestionDisplay      viewport.Model
-	// scrollbar              tea.Model
-}
-
-// type fileSuggestionInfo struct {
-// 	fileName    string
-// 	fileText    string
-// 	suggestions []suggestionState
-// }
-
-// type suggestionState struct {
-// 	isAccepted                                               bool
-// 	original, originalSuggestion, currentSuggestion, display string
-// }
-
-func newSuggestions(height *int) suggestions {
-	return suggestions{
-		height:                height,
-		currentFile:           "OEBS/Text/file.html",
-		currentSuggestionName: "Suggestion Name",
-	}
-}
-
-func (m suggestions) Init() tea.Cmd {
-	return nil
-}
-
-func (m suggestions) View() string {
-	return m.LeftStatusView()
-}
-
-func (m suggestions) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	return m, nil
-}
-
-func (m suggestions) LeftStatusView() string {
-	var (
-		statusView      = fmt.Sprintf("%s %s\n%s %s\n", documentIcon, fileNameStyle.Render(m.currentFile), suggestionIcon, suggestionNameStyle.Render(m.currentSuggestionName))
-		remainingHeight int
-		statusPadding   string
-	)
-
-	if m.height != nil {
-		remainingHeight = *m.height - (lipgloss.Height(statusView) + leftStatusBorderStyle.GetVerticalBorderSize())
-	}
-
-	if remainingHeight > 0 {
-		statusPadding = strings.Repeat("\n", remainingHeight)
-	}
-
-	return leftStatusBorderStyle.Render(statusView + statusPadding)
-}
