+++ testtui/sectionbreak.go
@@ -1,0 +1,58 @@
+package main
+
+import (
+	"strings"
+
+	"github.com/charmbracelet/bubbles/textinput"
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+type sectionBreak struct {
+	input        textinput.Model
+	contextBreak string
+}
+
+func newSectionBreak(focus bool) sectionBreak {
+	ti := textinput.New()
+	ti.Width = 20
+	ti.CharLimit = 200
+	ti.Placeholder = "Section break"
+
+	if focus {
+		ti.Focus()
+	}
+
+	return sectionBreak{
+		input: ti,
+	}
+}
+
+func (m sectionBreak) Init() tea.Cmd {
+	return nil
+}
+
+func (m sectionBreak) View() string {
+	return m.input.View()
+}
+
+func (m sectionBreak) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	var (
+		cmds []tea.Cmd
+		cmd  tea.Cmd
+	)
+	m.input, cmd = m.input.Update(msg)
+	cmds = append(cmds, cmd)
+
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch msg.String() {
+		case "enter":
+			m.contextBreak = strings.TrimSpace(m.input.Value())
+			if m.contextBreak != "" {
+				cmds = append(cmds, nextStage)
+			}
+		}
+	}
+
+	return m, tea.Batch(cmds...)
+}
